+++ Array/17EGICS048_ARRAY_2
