+++ Array/17EGICS055_ARRAY_2
