+++ Array/17EGICS071_Array_02.c
