+++ Array/17EGICS071_Array_2.py
