+++ Array/17EGICS083_Array_1.cpp
