+++ Array/17EGICS083_Array_2.cpp
