+++ Array/17EGICS092_ARRAY_1.py
