+++ Array/17EGICS300_ARRAY_1
