+++ Array/17egics017_ARRAY_2.py
